CESAR TRY/requirements.txt
git
transformers
llama-cpp-python

# CMake